requirements.txt
beautifulsoup4==4.12.2
selenium==4.18.1
pandas==2.2.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
webdriver-manager==4.0.1